requirement.txt
aniso8601==1.3.0
blinker==1.4
click==6.7
Flask==0.12.2
Flask-JWT==0.3.2
Flask-Mail==0.9.1
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.4.2
pymongo==3.5.1
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
Werkzeug==0.12.2
